> Renames clan pages
